pyproject.toml
[tool.poetry]
name = "AlphaVentureDiscordBot"
version = "0.1.0"
description = "Hacky Bot to plot a graph of the current stock levels and post it to Discord."

license = "MIT"

authors = [
    "Jing Kai Tan <contact@jingk.ai>"
]

readme = 'README.md'  # Markdown files are supported

repository = "https://github.com/xlanor"
homepage = "https://github.com/xlanor"

keywords = ['finance','stocks','discord']

[tool.poetry.dependencies]
python = "^3.6"  # Compatible python versions must be declared here
toml = "^0.9"
alpha_vantage = "^2.1"
pandas = "^0.23.4"
matplotlib = "^3.0"
"discord.py" = "^0.16.12"
arrow = "^0.12.1"


[tool.poetry.dev-dependencies]
black = "=18.9b0"
